refactor(orpc-error): flatten API and document From boundary
Drop the internal error shim that leaked orpc::error::error via the
facade re-export, and document why From<IOError> must live in orpc.
